+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.12.0-rc.35",
-  genVersion: "2.691.6",
-  userAgent: "speakeasy-sdk/typescript 3.12.0-rc.35 2.691.6 2.0 @orq-ai/node",
+  sdkVersion: "3.12.0-rc.36",
+  genVersion: "2.692.0",
+  userAgent: "speakeasy-sdk/typescript 3.12.0-rc.36 2.692.0 2.0 @orq-ai/node",
 } as const;