+++ packages/orq-rc/RELEASES.md
@@ -3409,3 +3409,13 @@ Based on:
 - [typescript v3.12.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v3.12.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.35 - packages/orq-rc
+
+## 2025-09-03 19:09:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.609.0 (2.692.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.12.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v3.12.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.36 - packages/orq-rc